MetricsManager is an interface
