+++ frontend/src/lib/SegmentMetricsManager.test.ts
